+++ src/common/components/Dropdown/index.vue
@@ -25,7 +25,7 @@
 
 <script>
 import { alertIcon } from "@/common/icons"
-import { validateInput } from "@/common/mixins"
+import { validateInput } from "@/common/lib/validate"
 
 export default {
   name: "UiDebioDropdown",
