+++ src/common/components/File/index.vue
@@ -16,7 +16,7 @@
 <script>
 import Button from "@/common/components/Button"
 import { alertIcon } from "@/common/icons"
-import { validateInput } from "@/common/mixins"
+import { validateInput } from "@/common/lib/validate"
 
 export default {
   name: "UiDebioFile",
